move upgrade container test
